moving tests outside the module
