+++ tests/trapezoid_test.py
@@ -1,68 +1,63 @@
-import sys
-
-sys.path.append("../")
-
-from integration.gaussian import GaussLegendre
+from torchquad.integration.trapezoid import Trapezoid
 from helper_functions import (
     compute_integration_test_errors,
     setup_test_for_backend,
 )
 
 
-def _run_gaussian_tests(backend, _precision):
-    """Test the integrate function in integration.gaussian for the given backend."""
+def _run_trapezoid_tests(backend, _precision):
+    """Test the integrate function in integration.Trapezoid for the given backend."""
 
-    gauss = GaussLegendre()
+    tp = Trapezoid()
 
     # 1D Tests
-    N = 60
-
+    N = 100000
     errors, funcs = compute_integration_test_errors(
-        gauss.integrate,
+        tp.integrate,
         {"N": N, "dim": 1},
         integration_dim=1,
         use_complex=True,
         backend=backend,
     )
-    print(f"1D {gauss} Test passed. N: {N}, backend: {backend}, Errors: {errors}")
-    # Polynomials up to degree 2N-1 can be integrated almost exactly with gaussian.
+    print(f"1D Trapezoid Test passed. N: {N}, backend: {backend}, Errors: {errors}")
+    # Polynomials up to degree 1 can be integrated almost exactly with Trapezoid.
     for err, test_function in zip(errors, funcs):
-        assert test_function.get_order() > (2 * N - 1) or err < 5.5e-11
+        assert test_function.get_order() > 1 or (
+            err < 2e-11 if test_function.is_integrand_1d else err < 5e-10
+        )  # errors add up if the integrand is higher dimensional
     for error in errors:
-        assert error < 6.33e-11
+        assert error < 1e-5
 
     N = 2  # integration points, here 2 for order check (2 points should lead to almost 0 err for low order polynomials)
-
     errors, funcs = compute_integration_test_errors(
-        gauss.integrate,
+        tp.integrate,
         {"N": N, "dim": 1},
         integration_dim=1,
         use_complex=True,
         backend=backend,
     )
-    print(f"1D {gauss} Test passed. N: {N}, backend: {backend}, Errors: {errors}")
-    # All polynomials up to degree 3 = 2N-1 should be 0, others should be good as well.
+    print(f"1D Trapezoid Test passed. N: {N}, backend: {backend}, Errors: {errors}")
+    # All polynomials up to degree = 1 should be 0
     # If this breaks check if test functions in helper_functions changed.
     for err, test_function in zip(errors, funcs):
-        assert test_function.get_order() > 3 or err < 8e-15
+        assert test_function.get_order() > 1 or err < 1e-15
     for error in errors[:2]:
         assert error < 1e-15
 
     # 3D Tests
-    N = 60**3
-
+    N = 1000000
     errors, funcs = compute_integration_test_errors(
-        gauss.integrate,
+        tp.integrate,
         {"N": N, "dim": 3},
         integration_dim=3,
         use_complex=True,
         backend=backend,
     )
-    print(f"3D {gauss} Test passed. N: {N}, backend: {backend}, Errors: {errors}")
+    print(f"3D Trapezoid Test passed. N: {N}, backend: {backend}, Errors: {errors}")
     for err, test_function in zip(errors, funcs):
         assert test_function.get_order() > 1 or (
-            err < 1e-12 if test_function.is_integrand_1d else err < 1e-11
-        )
+            err < 1e-12 if test_function.is_integrand_1d else err < 2e-11
+        )  # errors add up if the integrand is higher dimensional
     for error in errors:
         assert error < 6e-3
 
@@ -73,24 +68,23 @@ def _run_gaussian_tests(backend, _precision):
         return
 
     # 10D Tests
-    N = (60**3) * 3
-
+    N = 10000
     errors, funcs = compute_integration_test_errors(
-        gauss.integrate,
+        tp.integrate,
         {"N": N, "dim": 10},
         integration_dim=10,
         use_complex=True,
         backend=backend,
     )
-    print(f"10D {gauss} Test passed. N: {N}, backend: {backend}, Errors: {errors}")
+    print(f"10D Trapezoid Test passed. N: {N}, backend: {backend}, Errors: {errors}")
     for err, test_function in zip(errors, funcs):
-        assert test_function.get_order() > 60 or err < 4e-09  # poly order should be relatively high
+        assert test_function.get_order() > 1 or err < 1e-11
     for error in errors:
-        assert error < 1e-5
+        assert error < 7000
 
     # JIT Tests
     if backend != "numpy":
-        N = 60
+        N = 100000
         jit_integrate = None
 
         def integrate(*args, **kwargs):
@@ -98,7 +92,7 @@ def _run_gaussian_tests(backend, _precision):
             # which is then re-used on all other integrations (as is the point of JIT).
             nonlocal jit_integrate
             if jit_integrate is None:
-                jit_integrate = gauss.get_jit_compiled_integrate(dim=1, N=N, backend=backend)
+                jit_integrate = tp.get_jit_compiled_integrate(dim=1, N=N, backend=backend)
             return jit_integrate(*args, **kwargs)
 
         errors, funcs = compute_integration_test_errors(
@@ -110,12 +104,15 @@ def _run_gaussian_tests(backend, _precision):
             filter_test_functions=lambda x: x.is_integrand_1d,
         )
 
-        print(f"1D Gaussian JIT Test passed. N: {N}, backend: {backend}, Errors: {errors}")
-        # Polynomials up to degree 2N-1 can be integrated almost exactly with gaussian.
+        print(
+            f"1D Trapezoid JIT Test passed for 1D integrands. N: {N}, backend: {backend}, Errors: {errors}"
+        )
         for err, test_function in zip(errors, funcs):
-            assert test_function.get_order() > (2 * N - 1) or err < 2e-10
+            assert test_function.get_order() > 1 or (
+                err < 2e-11 if test_function.is_integrand_1d else err < 5e-10
+            )  # errors add up if the integrand is higher dimensional
         for error in errors:
-            assert error < 6.33e-11
+            assert error < 1e-5
 
         jit_integrate = None  # set to None again so can be re-used with new integrand shape
 
@@ -128,18 +125,20 @@ def _run_gaussian_tests(backend, _precision):
             filter_test_functions=lambda x: x.integrand_dims == [2, 2, 2],
         )
         print(
-            f"1D Gaussian JIT Test passed for [2, 2, 2] dimensional integrands. N: {N}, backend: {backend}, Errors: {errors}"
+            f"1D Trapezoid JIT Test passed for [2, 2, 2] dimensional integrands. N: {N}, backend: {backend}, Errors: {errors}"
         )
         for err, test_function in zip(errors, funcs):
-            assert test_function.get_order() > 1 or err < 2e-10
+            assert test_function.get_order() > 1 or (
+                err < 2e-11 if test_function.is_integrand_1d else err < 5e-10
+            )  # errors add up if the integrand is higher dimensional
         for error in errors:
             assert error < 1e-5
 
 
-test_integrate_numpy = setup_test_for_backend(_run_gaussian_tests, "numpy", "float64")
-test_integrate_torch = setup_test_for_backend(_run_gaussian_tests, "torch", "float64")
-test_integrate_jax = setup_test_for_backend(_run_gaussian_tests, "jax", "float64")
-test_integrate_tensorflow = setup_test_for_backend(_run_gaussian_tests, "tensorflow", "float64")
+test_integrate_numpy = setup_test_for_backend(_run_trapezoid_tests, "numpy", "float64")
+test_integrate_torch = setup_test_for_backend(_run_trapezoid_tests, "torch", "float64")
+test_integrate_jax = setup_test_for_backend(_run_trapezoid_tests, "jax", "float64")
+test_integrate_tensorflow = setup_test_for_backend(_run_trapezoid_tests, "tensorflow", "float64")
 
 
 if __name__ == "__main__":
